Add PUT,DELETE exception handling

+++ src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -47,15 +47,21 @@ public class UserController {
 
     @PutMapping("/user")
     public void updateUser(@RequestBody UserUpdateRequest request) {
-        String sql = "UPDATE user SET name = ? WHERE id = ?";
+        String readSql = "SELECT * FROM user WHERE id = ?";
+        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, request.getId()).isEmpty();
+        if (isUserNotExist) throw new IllegalArgumentException();
 
+        String sql = "UPDATE user SET name = ? WHERE id = ?";
         jdbcTemplate.update(sql, request.getName(), request.getId());
     }
 
     @DeleteMapping("/user")
     public void deleteUser(@RequestParam String name) {
-        String sql = "DELETE FROM user WHERE name = ?";
+        String readSql = "SELECT * FROM user WHERE name = ?";
+        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, name).isEmpty();
+        if (isUserNotExist) throw new IllegalArgumentException();
 
+        String sql = "DELETE FROM user WHERE name = ?";
         jdbcTemplate.update(sql, name);
     }
